requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
gitpython
chromadb
python-dotenv
langchain-huggingface
torch